gamepad work

+++ src/firmware_esp32/src/stepper.h
@@ -12,6 +12,6 @@ extern ShiftRegister74HC595<NUMBER_SHIFT_REG> sr;
 
 void do_half_step(char PIN_STEPPER_X_STEP, int counter);
 void set_direction(char PIN_STEPPER_X_STEP, float sign);
-void go_to_next_set_point(float delta_q1, float delta_q2, float delta_q3, float delta_t);
+void go_to_next_set_point(float delta_q1, float delta_q2, float delta_q3, long unsigned int delta_t);
 
 #endif // STEPPER_H
